🐛 FIX: 현재 인원 필드에 대한 동시성 문제 해결
- 현재 참가 인원에 대한 동시성 문제를 해결하기 위해 AtomicInteger 타입으로 변경했습니다.
- 참가 인원 증가 메서드에 CAS 기법을 적용하여 성능 문제 없이 동시성 문제를 해결하고자 합니다.

+++ src/main/java/cloudcomputinginha/demo/converter/InterviewConverter.java
@@ -79,7 +79,7 @@ public class InterviewConverter {
                 .sessionName(interview.getSessionName())
                 .jobName(interview.getJobName())
                 .interviewType(interview.getInterviewOption().getInterviewType())
-                .currentParticipants(interview.getCurrentParticipants())
+                .currentParticipants(interview.getCurrentParticipants().intValue())
                 .maxParticipants(interview.getMaxParticipants())
                 .startedAt(interview.getStartedAt())
                 .build();
@@ -125,7 +125,7 @@ public class InterviewConverter {
                 .jobName(interview.getJobName())
                 .interviewType(interview.getInterviewOption().getInterviewType())
                 .maxParticipants(interview.getMaxParticipants())
-                .currentParticipants(interview.getCurrentParticipants())
+                .currentParticipants(interview.getCurrentParticipants().intValue())
                 .startedAt(interview.getStartedAt())
                 .hostName(
                         memberInterviewList.stream()
@@ -163,7 +163,7 @@ public class InterviewConverter {
                 .name(interview.getName())
                 .corporateName(interview.getCorporateName())
                 .jobName(interview.getJobName())
-                .currentParticipants(interview.getCurrentParticipants())
+                .currentParticipants(interview.getCurrentParticipants().intValue())
                 .maxParticipants(interview.getMaxParticipants())
                 .startedAt(interview.getStartedAt())
                 .endedAt(interview.getEndedAt())
